requirements.txt
discord[voice]~=2.1.0
pynacl~=1.5.0
peewee~=3.15.4
Pillow~=9.4.0
numpy~=1.24.1
requests~=2.28.2
pymysql~=1.0.2
cryptography~=39.0.0
